requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.0
opencv-python==4.8.0.76
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
tqdm==4.66.1
transformers>=4.40.0
